pyproject.toml
[project]
name = "chroma-knowledge-search"
version = "0.1.0"
description = ""
authors = [
    {name = "Jah-Wilson Teeba",email = "therealjwil07@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.120.4,<0.121.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "openai (>=2.6.1,<3.0.0)",
    "chromadb (>=1.3.0,<2.0.0)",
    "pdfminer-six (>=20250506,<20250507)",
    "python-docx (>=1.2.0,<2.0.0)",
    "tenacity (>=9.1.2,<10.0.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "bcrypt (>=5.0.0,<6.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pydantic (>=2.12.3,<3.0.0)",
    "pytest (>=8.4.2,<9.0.0)",
    "pytest-asyncio (>=0.21.0,<1.0.0)",
    "pytest-cov (>=4.0.0,<5.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "itsdangerous (>=2.2.0,<3.0.0)",
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "streamlit (>=1.51.0,<2.0.0)",
    "selenium (>=4.15.0,<5.0.0)",
    "pytest-mock (>=3.12.0,<4.0.0)",
    "httpx (>=0.28.1,<0.29.0)"
]

[tool.poetry]
packages = [{include = "chroma_knowledge_search", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
